Minor PHPUnit related reorganisation.
The phpunit.xml file is only relevant for developers of the
beeminder-php-api library, so I have moved it into the test directory so
that users of the library don't need to think about it at all.

Updated the README with information about test coverage and how to run
the tests.